+++ app/Main.hs
@@ -7,12 +7,12 @@
 module Main where
 
 import Options.Generic
-import Data.Text (unpack)
-import System.IO (stderr, stdout, Handle, IOMode(..), openFile, hClose)
+import System.IO (stderr, stdout, Handle, IOMode(..), openFile, hClose, hPutStrLn)
 import System.FilePath (takeDirectory, (</>))
 import System.Directory (setCurrentDirectory, getCurrentDirectory)
 import qualified Data.Text.IO as T
 import Data.FileEmbed
+import System.Exit (exitFailure)
 
 import Celtchar.Novel.Structure
 import Celtchar.Novel
@@ -34,7 +34,6 @@ main = do
     cmd <- getRecord "celtchar" :: IO Command
 
     let conf = root cmd
-    h <- getOutputHandle $ output cmd
     f <- getNovelStructure $ conf
 
     let inDir = takeDirectory $ root cmd
@@ -44,13 +43,15 @@ main = do
     -- write the final tex file
     setCurrentDirectory inDir
 
-    case f of Just x  -> do res <- stringify (language x) (novelify x)
-                            T.hPutStr h res
-              Nothing -> T.hPutStrLn stderr "error while parsing"
+    case f of Right x  -> do
+                res <- stringify (language x) (novelify x)
 
-    hClose h
+                setCurrentDirectory rootDir
+                h <- getOutputHandle $ output cmd
+                T.hPutStr h res
+                hClose h
 
-    -- write the sty file
-    setCurrentDirectory rootDir
-
-    T.writeFile (outDir </> "ogma.sty") $(embedStringFile "ogma.sty")
+                T.writeFile (outDir </> "ogma.sty") $(embedStringFile "ogma.sty")
+              Left err -> do
+                hPutStrLn stderr err
+                exitFailure