+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # celtchar
 
+## celtchar-0.2.0.0
+
+### Feature
+
+* If `celtchar` cannot parse the configuration file, it reports an error
+
 ## celtchar-0.1.0.1
+
+### Bugfix
 
 * Fix the cabal file to list the `ogma.sty` file